requirements.txt
Flask>=2.2,<3
pycryptodome
python-dotenv
flask-cors
